Refactor (1 type)
