--- src/requirements/Yup/components/Score.tsx
@@ -1,8 +1,0 @@
-import { RequirementFormProps } from "requirements"
-import MinScoreField from "./MinScoreField"
-
-const Score = ({ baseFieldPath }: RequirementFormProps): JSX.Element => (
-  <MinScoreField baseFieldPath={baseFieldPath} />
-)
-
-export default Score
